Fix: Address various linter warnings (#90)
* Refactor: Address various linter warnings

This commit addresses a series of linter warnings reported by clang-tidy to improve code quality, readability, and maintainability.

The following changes have been made:

- Suppressed `bugprone-easily-swappable-parameters` warnings for callback functions (`compute_vector`, `hashing_worker_callback`) and a test helper (`create_test_file`) using `NOLINTNEXTLINE` comments, as altering their signatures was impractical.
- In `src/human_size.c`, replaced the magic number `1024.0` with the named constant `BYTES_IN_KIBIBYTE` and renamed the loop counter `i` to `unit_index` for better clarity.
- In `src/napr_threadpool.c`, renamed the short loop variable `l` to `idx` to improve readability in the threadpool initialization and shutdown functions.

* Fix: Address various linter warnings

This commit addresses a number of clang-tidy warnings to improve code quality and ensure CI checks pass.

- Suppressed `bugprone-easily-swappable-parameters` in a test utility.
- Fixed `bugprone-narrowing-conversions` with explicit casts.
- Fixed `cert-err33-c` by casting `fprintf` return values to `(void)`.
- Fixed `readability-else-after-return` by removing unnecessary `else` blocks.
- Fixed `readability-braces-around-statements` by adding braces to single-line `if` statements.
- Fixed `bugprone-multi-level-implicit-pointer-conversion` with explicit casts.
- Replaced magic numbers with the `ERR_BUF_SIZE` constant.

---------

+++ src/human_size.c
@@ -67,5 +67,5 @@ apr_off_t parse_human_size(const char *size_str)
         }
     }
 
-    return (apr_off_t) (size * multiplier);
+    return (apr_off_t) (size * (double) multiplier);
 }
